requirements.txt
rich
python-dotenv
google-generativeai
